Added parallel routing in admin dashboard
