+++ src/app/admin/dashboard/@participants/default.tsx
@@ -1,0 +1,3 @@
+export default function ParticipantsDefault() {
+  return null;
+}
