+++ src/app/admin/dashboard/@participants/loading.tsx
@@ -1,0 +1,8 @@
+export default function ParticipantsLoading() {
+  return (
+    <div className="space-y-4">
+      <div className="h-8 bg-gray-200 rounded animate-pulse" />
+      <div className="h-64 bg-gray-200 rounded animate-pulse" />
+    </div>
+  );
+}
