+++ src/app/admin/dashboard/@workshops/default.tsx
@@ -1,0 +1,3 @@
+export default function WorkshopsDefault() {
+  return null;
+}
