+++ src/app/admin/dashboard/@workshops/loading.tsx
@@ -1,0 +1,8 @@
+export default function WorkshopsLoading() {
+  return (
+    <div className="space-y-4">
+      <div className="h-8 bg-gray-200 rounded animate-pulse" />
+      <div className="h-64 bg-gray-200 rounded animate-pulse" />
+    </div>
+  );
+}
